move PublishingProps alongside Publisher and alias ConfigureNodePluginProps
